+++ packages/flame_behaviors/lib/src/behaviors/events/hoverable_behavior.dart
@@ -1,4 +1,4 @@
-import 'package:flame/components.dart';
+import 'package:flame/events.dart';
 import 'package:flame_behaviors/flame_behaviors.dart';
 
 /// {@template hoverable_behavior}
@@ -9,4 +9,11 @@ import 'package:flame_behaviors/flame_behaviors.dart';
 /// behaviors.
 /// {@endtemplate}
 abstract class HoverableBehavior<Parent extends EntityMixin>
-    extends Behavior<Parent> with Hoverable {}
+    extends Behavior<Parent> with HoverCallbacks {
+  /// {@macro hoverable_behavior}
+  HoverableBehavior({
+    super.children,
+    super.priority,
+    super.key,
+  });
+}