+++ packages/flame_behaviors/test/helpers/helpers.dart
@@ -1,2 +1,1 @@
-export 'test_game.dart';
 export 'throws_behavior_not_found_for.dart';