+++ packages/flame_behaviors/test/helpers/test_game.dart
@@ -1,4 +1,0 @@
-import 'package:flame/game.dart';
-
-class TestGame extends FlameGame
-    with HasCollisionDetection, HasDraggables, HasHoverables, HasTappables {}